move superglobals to header.php

+++ deletenote.php
@@ -1,23 +1,25 @@
 <?php
 include 'header.php';
 
-//  GET PARAMETERS
-$title_input = $_GET['title'];
-
-// QUERY EXECUTION
-try {
-    //  DELETE
-    $sql = "DELETE FROM $tbname WHERE title = '$title_input'";
-    if ($sql) {
-        $result = $pdo->query($sql);
-        $feedback['succes'] = $title_input . " was succesfully deleted";
-        unset($result);
-    } else {
-        $feedback['querryError'] = "No records matching your query were found.";
+//  VALIDATE
+if (empty($title_input)) {
+    $feedback['validate_titleError'] = "please fill in a title.";
+} else {
+    // QUERY EXECUTION
+    try {
+        //  DELETE
+        $sql = "DELETE FROM $tbname WHERE title = '$title_input'";
+        if ($sql) {
+            $result = $pdo->query($sql);
+            $feedback['succes'] = $title_input . " was succesfully deleted";
+            unset($result);
+        } else {
+            $feedback['querryError'] = "No records matching your query were found.";
+        }
+    } catch (PDOException $e) {
+        $feedback['sqlError'] = $e->getMessage();
     }
-} catch (PDOException $e) {
-    $feedback['sqlError'] = $e->getMessage();
 }
 
-include 'feedback.php';
 
+include 'feedback.php';
